Fix #158 : Added SX1276 related functionalities to android repo (#162)
* Fix #158 : Added SX1276 related functionalities to android repo

* Reviewed changes fixed!

+++ app/src/main/java/org/fossasia/pslab/communication/peripherals/SPI.java
@@ -94,10 +94,10 @@ public class SPI {
         packetHandler.sendInt(value);
     }
 
-    public ArrayList<Byte> xfer(int channel, ArrayList<Integer> data) throws IOException {
+    public ArrayList<Byte> xfer(int channel, ArrayList<Byte> data) throws IOException {
         start(channel);
         ArrayList<Byte> reply = new ArrayList<>();
-        for (Integer a : data) {
+        for (Byte a : data) {
             reply.add(send8(a));
         }
         stop(channel);
